backend/pyproject.toml
[project]
name = "physical-ai-backend"
version = "0.1.0"
description = "Backend for Physical AI Textbook RAG Chatbot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "qdrant-client>=1.7.0",
    "httpx>=0.26.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "beautifulsoup4>=4.12.0",
    "tiktoken>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
